refactor app logic into viewmodel
